+++ CHANGELOG.md
@@ -17,7 +17,7 @@
 - Zhinst backend - Raise a more understandable exception when compiling an acquisition with larger than allowed duration (!407).
 - Gettables - Added a ProfiledScheduleGettable for profiling execution times of schedule steps. (!420, !469)
 - Instrument Coordinator - Small fix for `search_settable_param` when scheduler is searching for qcodes parameters (!461)
-
+- Changed hyphens into underscores in order for DeviceElement names to comply with qcodes versions 0.34 and up. (#300)
 ## 0.8.0 Support for two qubit operations and basic CZ-gate implementation (2022-08-10)
 
 ### Breaking changes